Refactor unused imports and comments in auth and common modules
